# api-docs/swagger.yaml
basePath: /cosmos
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      status:
        example: OK
        type: string
    type: object
...
  utils.HTTPErrorResult:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API for Cosmos Cloud server management
  title: Cosmos Server API
  version: 0.23.00-unstable001
paths:
  /_logs:
    get:
      description: Returns the server log file as a plain text attachment
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Download server logs
      tags:
      - system
  /api/_memory:
    get:
      description: Returns memory usage details for internal caches and buffers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Get internal memory status
      tags:
      - system
...
  /api/can-send-email:
    get:
      description: Returns whether the server has email sending capabilities configured
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      summary: Check if email sending is enabled
      tags:
      - system
...
  /api/force-server-update:
    post:
      description: Triggers a manual check for available server updates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Force check for server updates
      tags:
      - system
...
  /api/restart-server:
    get:
      description: Triggers a restart of the host machine (not available when running
        inside a container)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      security:
      - BearerAuth: []
      summary: Restart the host machine
      tags:
      - system
...
  /api/status:
    get:
      description: Returns the current server status including configuration, resource
        info, and system state
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/utils.HTTPErrorResult'
      summary: Get server status
      tags:
      - system
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"